Refactor logs customization to defer prompt_toolkit imports.
